% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_server.R
\name{sdm_server}
\alias{sdm_server}
\title{Sistema de Dados Mestre (SDM) server address}
\usage{
sdm_server()
}
\value{
A string containing the SDM server URL.
}
\description{
Sistema de Dados Mestre (SDM) server address
}
